# versions/before-backend-bugfix-20260922-133652/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "BookingConfigOut": {
        "properties": {
          "campus_key": {
            "title": "Campus Key",
            "type": "string"
          },
          "external_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Url"
          },
          "line_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Url"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "mode": {
            "$ref": "#/components/schemas/BookingMode"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "campus_key",
          "mode",
          "version",
          "line_url",
          "phone",
          "external_url",
          "message"
        ],
        "title": "BookingConfigOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "VisitSlotUpdateRequest": {
        "properties": {
          "capacity": {
            "anyOf": [
              {
                "maximum": 200.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "closed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed"
          }
        },
        "title": "VisitSlotUpdateRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Ivy Website Admin API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/website/v1/admin/booking-config/{campus_key}": {
      "get": {
        "operationId": "get_booking_config_api_website_v1_admin_booking_config__campus_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "campus_key",
            "required": true,
            "schema": {
              "title": "Campus Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "ivy_admin_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ivy Admin Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingConfigOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Booking Config",
        "tags": [
          "booking"
        ]
      },
      "patch": {
        "operationId": "update_booking_config_api_website_v1_admin_booking_config__campus_key__patch",
        "parameters": [
          {
            "in": "path",
            "name": "campus_key",
            "required": true,
            "schema": {
              "title": "Campus Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "ivy_admin_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ivy Admin Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingConfigUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingConfigOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Booking Config",
        "tags": [
          "booking"
        ]
      }
    },
...
    "/api/website/v1/admin/slots": {
      "get": {
        "operationId": "list_admin_slots_api_website_v1_admin_slots_get",
        "parameters": [
          {
            "in": "query",
            "name": "campus_key",
            "required": true,
            "schema": {
              "title": "Campus Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Date From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Date To",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "ivy_admin_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ivy Admin Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VisitSlotOut"
                  },
                  "title": "Response List Admin Slots Api Website V1 Admin Slots Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Admin Slots",
        "tags": [
          "booking"
        ]
      },
      "post": {
        "operationId": "create_admin_slot_api_website_v1_admin_slots_post",
        "parameters": [
          {
            "in": "query",
            "name": "campus_key",
            "required": true,
            "schema": {
              "title": "Campus Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "ivy_admin_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ivy Admin Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisitSlotCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitSlotOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Admin Slot",
        "tags": [
          "booking"
        ]
      }
    },
    "/api/website/v1/admin/slots/{slot_id}": {
      "patch": {
        "operationId": "update_admin_slot_api_website_v1_admin_slots__slot_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "slot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Slot Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "ivy_admin_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ivy Admin Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisitSlotUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitSlotOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Admin Slot",
        "tags": [
          "booking"
        ]
      }
    },
...
    "/api/website/v1/admin/visit-requests/export": {
      "get": {
        "operationId": "export_visit_requests_api_website_v1_admin_visit_requests_export_get",
        "parameters": [
          {
            "in": "query",
            "name": "campus_key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campus Key"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "ivy_admin_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ivy Admin Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Visit Requests",
        "tags": [
          "booking"
        ]
      }
    },
...
  }
}